Начинает рефакторинг
